Make Skia corpus tests pass for Mac ARM64 with macOS 15
Add a bunch of failing corpus tests to SUPPRESSIONS_EXACT_MATCHING so
they can pass with fuzzy matching. For the tests that cannot pass,
update the test expectations to use the ARM64 expectations and suppress
the tests for Mac Intel, by rolling DEPS for the testing corpus to pick
up https://pdfium-review.googlesource.com/138030.

Bug: 41480180,459586268
Change-Id: Id9d3fb51a88032d3b39d3fa5f8a6fad80a9bdfeb