[Skia] Cleanup a complex reference lifetime
The old code was written in a way that looks like it was intended to
prevent a copy of stroke_options but due to the complicated flow chart
of ternary operator behavior it was actually copied. The old code was in
fact safe because the ternary operator resulted in a prvalue which
activates const ref lifetime extension, however that is a complicated
feature to rely on for memory safety.

The new code will result in the same one-time copy of stroke_options but
now that is explictly written. Now readers don't have to look up complex
reference lifetime rules to determine that this code is correct.

Bug: 334774307
Change-Id: I53a9cd8db8b87b748719079e93f3ab07414fad9c