[Skia] Add a new pixel text for single point paths
This test case exercises many different ways to make a single point
path.

PDFium currently renders some of the dots when the spec says not to.
Only the center ink dot, the 4 blue dots, and the 2 magenta dots should
be rendered. The square cap green dots should render nothing.

This is defined in Section 8.5.3.2 of ISO 32000-1:2008 (the last
paragraph).

Bug: 334774307, 460180742
Change-Id: Iaa773411a8b5a7d05cd50a2ef167ad5a98a7b9f4