[Skia] Add a new pixel test for single point paths
This test case exercises many different ways to make a single point
path.

Skia currently renders some of the dots when the spec says not to.
Only the center ink dot, the 4 blue dots, and the 2 magenta dots should
be rendered. The square cap green dots should render nothing. Also the
"not generally useful" single device pixel fills are not currently
painted by the Skia renderer.

All of the renderers have different issues with this test. Details in
the bug.

Stroke painting is defined in Section 8.5.3.2 of ISO 32000-1:2008 (the
last paragraph). Fill painting is defined in Section 8.5.3.3.1. Path
strokes and fills forming an implicit knockout group is defined in
section 11.7.4.4.

Bug: 334774307, 460180742
Change-Id: Iaa773411a8b5a7d05cd50a2ef167ad5a98a7b9f4