Update CFX_Face::Open() Callers in skiafontmgr
This CL creates a new function CFX_Face:::OpenFromFilePath() which
constructs ft_open_args which is called in skiafontmgr. This is done to
avoid arg construction in the Callee and is a continuation of the effort
to consolidate CFX_Face::Open() Callers and eventually make the previously
mentioned method private.

Bug: 459848480
Change-Id: Ieeeccb66342a34b6c5826927ce6016449e93dcca