Make GetRec() private with ScopedFontTransform as a friend class
This CL makes GetRec() private, with the addition of ScopedFontTransform
as a friend class. This is done to avoid exposing struct FT_FaceRec_.
This was done after previous consolidation of other public callers and
movement of the core logic in these callers to cfx_face.cpp. Allowing
GetRec() to be private.

Bug: 460068801
Change-Id: I6e16c65f1452e803f18f93aa73e7bbe8af867bc4