Decouple anti-alias setting from FT_Render_Mode
This CL decouples the anti-alias setting from the previous freetype
callers. This is done by introducing a new enum, AARenderMode, and a
function GetFTRenderMode(), which is used to abstract the antialiasing
setting.

This change is part of the ongoing effort to encapsulate FreeType usage,
which makes future adjustments to anti-aliasing logic easier and
cleaner.

Change-Id: I6ef3d17ca6af94972e48691311df088f5cb6f64d
Bug: 452087737