Unified GetFooParamValue() and SetFooParam() implementations
This CL aims to have a common function, which is called for each call of
GetFooParamValue() and SetFooParam() functions that are present in
fpdfsdk/fpdf_editpage.cpp. This CL takes advantage of the common
behaviour between these functions to unify them. This is done through
two functions, SetParamValueHelper() and GetParamValueHelper(). These
two functions take over setting pParams and pObj for all variable types,
reducing common code.

This unification aims to reduce code duplication among these functions,
which can make the section both shorter and easier to edit in the
future.

Bug: 372062063
Change-Id: I74657690738e440cd483c2ccc3d9e0f847bd6846