Improve `pdfium_test --png` implementation
Internally, pdfium_test has a WritePng() function that calls
EncodePng() with the bitmap format hard-coded to `FPDFBitmap_BGRA`, even
if the bitmap format is actually `FPDFBitmap_BGRx`. Add a bool to
struct BitmapAttributes to indicate whether the page has an alpha value.
Then use it in WritePng() to select the bitmap format appropriately.

Change-Id: Ib439fe0afbfd282e568b6f54491a2ef9ba290d9f