Connect public FPDF_FONT_BACKEND_TYPE with FontLibrary::SetType()
Do the plumbing work so the PDFium embedder's preference between
FreeType and Fontations gets passed to FontLibrary when initializing
PDFium. Then replace direct calls to create FreeTypeFontLibrary with
calls to FontLibrary::Create().

When the selected font library is FreeType, this is a no-op. When the
selected font library is Fontations, just crash deliberately for now,
as Fontations support has not been implemented yet.

Bug: 42271123
Change-Id: I4199211c73a917b835c769934a3c740a23614cd9