Split CFX_Face::GetFontStyleInfo() into 2 to avoid returning a struct
CL [1] added GetFontStyleInfo() to return a struct, but it is actually
slightly simpler to just have GetFontStyle() that returns the first part
of the struct. Callers that need the second part of the struct can call
CFX_Face::GetOs2CodePageRange() directly. This results in 2
GetOs2CodePageRange() calls, since GetFontStyle() calls it internally as
well, but it really is not that expensive to do.

Along the way:

- In core/fxge/android/cfpf_skiafontmgr.cpp, fix an incorrect code page
  range refactor from [1].
- Mark GetFontStyle() and GetOs2CodePageRange() as XFA/Android-only.
- Group all the conditional methods together in cfx_face.h.

[1] https://pdfium-review.googlesource.com/137430

Change-Id: I14b9959912feb2c777dc22a80e6cfdfc8d76f2ec