Add text extraction tests for zero-width characters
This CL aims to further explore how the character width check in
cpdf_textpage.cpp. The line at hand is below.

if (fabs(pTextObj->GetRect().Width()) < kSizeEpsilon)

This Cl aims to create new tests in fpdf_text_embeddertest.cpp to
exercise this specific case and examine how this rendering of text,
through the addition of tests which may have this as a root cause. This
sets up a foundation to remove the specific check in the future and get
a clearer picture of the impact of removing this check on rendering.

Bug: 444176962, 40643656
Change-Id: I99a6a5f0c4ccd1825437482845b597272f297d2f