Remove the CharWidth check
This patch removes the check for the width of the character in
cpdf_textpage.cpp. This prevents spaces (and some other thin characters)
from being rendered. The removal of this check ensures that these
characters are once again correctly rendered.

This patch also adds new tests and updates old NeedsImprovement results
for tests which have been fixed by the removal of this afformentioned
check.

Change-Id: I99a6a5f0c4ccd1825437482845b597272f297d2f
Bug: 444176962, 40643656