Partially fix CPDFToUnicodeMapTest.NonBmpUnicodeLookup
Make the lookup work on many POSIX platforms where wchar_t is 32-bit.

Start by getting rid of GetMultiCharIndexIndicator() and
`multi_char_vec_`. The synthesized values GetMultiCharIndexIndicator()
generates are problematic because they may overlap with values in
charmaps. Instead, change `map_` to store the lookup result directly,
instead of potentially doing a secondary lookup in `multi_char_vec_`.
Update InsertIntoMaps() to keep in sync with the data changes.

For the reverse mapping, change InsertIntoMaps() to see if `destcode`
is a WideString. If so, use SurrogatePair to decode the WideString and
get the code point. Use the code point as the key in `reverse_map_`.

Bug: 374947848
Change-Id: Ia8693184e0eb8408dd60aaebb4141181aff1c5cb