Make CPDF_ToUnicodeMap::ReverseLookup() work better on Windows
ReverseLookup() takes a wchar_t, and wchar_t cannot be used for a
Non-bmp Unicode reverse lookup when wchar_t is only 16-bits on Windows.
Fix this by switching to char32_t. Update affected and related code to
move away from wchar_t when appropriate, or cast to wchar_t when needed.

Bug: 374947848
Change-Id: I38d2adc6cde93b7a23a6f48ecaf69429d47139fc