Handle multiple plus and minus signs in string to num conversions
https://pdfium-review.googlesource.com/c/pdfium/+/124412 caused a
regression where strings with multiple leading plus and/or minus signs
would return a value of 0. While the strings could be considered
invalid input, there are already PDFs in the wild that have these
strings.

Handle them by skipping all plus and minus signs (except the last minus
sign). This does not completely revert the behavior prior to the
previous CL, but should be able to handle the only known PDF example
with this case.

Bug: 432208037
Change-Id: I43d344c9fd6082595104bfe5ba16dc927f599b66