Replace checksums with PNG image expectations for rectangles
This CL replaces the use of MD5 checksums with PNG image expectations
for the rectangles.pdf and many_rectangles.pdf test files. This change
uses the CompareBitmapToPngWithExpectationSuffix() function and
introduces new functions to test with PNG expectations.

This improves the clarity of these comparisons allowing the addition of
visual verification, which is more robust than the prior MD5 checksum.

The change also adds new PNG Comparison test helper functions such as
TestRenderPageBitmapWithMatrixToPng(),
TestRenderPageBitmapWithExternalMemoryAndNoStrideToPng() and many more.

Bug: 468228360
Change-Id: I8b7dfc61f2db08cad5c43d56bbd9ee61de9d6b3c