Delete CJBig2_Image::GetLine()
This safer version of GetLineUnsafe() is still not that safe, as it does
not return a span. Get rid of it and change production callers to use
GetLineUnsafe() instead. The production callers now must check the line
number is valid. Add a IsValidLine() helper method, split off from
IsValidPixel(), to do that. For test callers, also convert to
GetLineUnsafe(), or delete them if they are deliberately unsafe.

Change-Id: I1a442621bc6d51695b2a8ea1bf61a822160c14b9