Update fpdf_editimg_embeddertest.cpp to use PNGs instead of Checksums
This CL aims to utilize PNGs instead of md5 checksums in the
fpdf_editimg_embeddertest.cpp file. This was done by adding another
conversion method for BGR to BGRA to support CompareBitmapToPng() and
updating the file to utilize the function CompareBitmapToPng() with
images rather than the MD5 checksums. This makes rendering more visually
evident and robust than md5 checksums.

Bug: 468228360
Change-Id: I6ee44f5845aecb2cb170a131569f89ec64c90197