Update checksums to images for rectangles_multi_page
This CL replaces checksums with images (primarily in
fpdf_doc_embeddertest.cpp) but also aims to remove
rectangles_multi_pages checksum usage from other files as well. The use
of a PNG image instead of an MD5 checksum allows for a visual
understanding of these checksums and also is more robust as opposed to
raw comparison with just checksums.

Bug: 468228360
Change-Id: Iaec487b9293943480ecec2c5607674a988daeb20