Replace checksums with images in fpdf_editpage_embeddertest.cpp
This CL replaces the md5 hashes present in
fpdf_editpage_embeddertest.cpp with png images to be able to more
clearly see rendering visually. This method is more robust than the
previously used MD5 checksums and aims to achieve the goals of moving
all CompareBitmap calls to PNGs instead of checksums.

Bug: 468228360
Change-Id: If415483a0443e67f8f31878a537f76ae4421ddd9