Stop setting SBSYMCODELEN in REFAGGNINST > 1 arith case
CJBig2_TRDProc::DecodeArith() only reads SBSYMCODELEN if pIDS
is nullptr, but CJBig2_SDDProc::DecodeArith() always calls it with
a non-nullptr pIDS. That's a good thing too, since it computes
SBSYMCODELEN incorrectly. (`SBSYMCODELENA`, computed at the top
of the function, used to initialize the IAID in the passed-in
pIDS, uses the correct computation.)

No behavior change.

See also remark 2 on the linked bug.

Bug: 472139308
Change-Id: Ida6be63b88f66761c5da1608428af8e0c3ec1ca9